// src/theme/enum.ts
export enum Pallete {
  Eggplant = '#401263',
  PinkHopbrush = '#5C198D',
  fadedpink = '#FFADED',
  lightPink = '#FFD6F6',
  Whitishpink = '#9D75BB',
  pastaGreen = '#A0C870',
  GoldenGlow = '#FFDE91',
  lightBrown = '#FFE9E3',
  orangeMelon = '#FF8A7A',
  EbonyGray = '#35293E',
  whiteBackground = '#FFFFFF',
  backgroundPink = '#FFF1FC',
  secondary = '#FF994F',
  disabledPrimary = '#CEBADD',
  backgrooundPrimary = '#FAF2FF',
  cardSelected = '#FFF5EE',
  cardBorder = '#FFD6B9',
  backgroundSecondary = '#7D47A4',
  inputBackground = '#663D20',
  choosePathBg = '#FFEBDC',
  settipUpexperienceBg = '#BC51A5',
  black = '#0B0B0B',
  white = '#F0F2F3',
  heading = '#0F0412',
  placeHolderText = '#878289',
  formSecondaryHeading = '#3F3641',
  selectionTextColor = '#663D20',
  settipUpCardBg = '#FFC9F3',
  homeBgColor = '#F9F4FC',
  homeLineraGradientLight = '#140428',
  homeLinearGradientDark = '#772AAC',
  plainWhite = '#FFFFFF',
  darkBlack = '#000',
  cardBgLightPink = '#FFF2FC',
  pinkTextColor = '#FF76E1',
  linearGradientDark = '#4C0F79',
  linearGradientMedium = '#5C1989',
  linearGradientLight = '#8933A3',
  shadowColor = '#c3c3c3',

  PrimaryDarkest = '#35293E',
  PrimaryDarker = '#3E2664',
  Primary = '#5C2D84',
  PrimaryLighter = '#9976B2',
  SecondaryDarker = '#D97FB1',
  Secondary = '#E6ABCD',
  SecondaryLighter = '#F1D4E6',
  SecondaryLightest = '#FDF0F5',
  TertiaryGreen = '#9DC56D',
  TertiaryRed = '#F4887B',
  TertiaryYellow = '#FFDA8D',
  TertiaryBeige = '#FAE7D9',
  Inherit = 'inherit',
  Current = 'currentColor',
  Transparent = 'transparent',
  Black = 'rgb(0, 0, 0)',
  White = 'rgb(255, 255, 255)',
  ScarletGum = '#4A1471',
  PastelPink = '#FFD6DD',
  Biloba = '#CFABF0',
  BilobaViolet = '#C5ABF0',
  Green = '#04CC00',
}

export enum fonts {
  PrimaryJakartaRegular = 'PlusJakartaSans-Light',
  PrimaryJakartaMedium = 'PlusJakartaSans-Medium',
  PrimaryJakartaBold = 'PlusJakartaSans-Bold',
  PrimaryJakartaExtraBold = 'PlusJakartaSans-ExtraBold',
  PrimaryJakartaSemiBold = 'PlusJakartaSans-Regular',
  SecondaryDMSansRegular = 'DMSans-Regular',
  SecondaryDMSansMedium = 'DMSans-Medium',
  SecondaryDMSansBold = 'DMSans-Bold',
}